requirements.txt
openai
streamlit
pandas
requests
python-dotenv
gemini

